Reorganize resources based on their URLs
